backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
python-telegram-bot
requests
python-dotenv
python-multipart